Complete dataset and reviews vol2

+++ Data/Models/Coordinate.cs
@@ -1,9 +1,9 @@
 ﻿namespace Data.Models; 
 
 public class Coordinate {
-    public int Id { get; set; }
+    public long Id { get; set; }
     
-    public int AnnotationId { get; set; }
+    public long AnnotationId { get; set; }
     public Annotation Annotation { get; set; }
 
     public int X { get; set; }
